+++ uplink/src/collector/tcpjson.rs
@@ -10,11 +10,11 @@ use tokio::{select, time};
 use tokio_stream::StreamExt;
 use tokio_util::codec::{Framed, LinesCodec, LinesCodecError};
 
-use std::{collections::HashMap, io, sync::Arc};
 use std::pin::Pin;
+use std::{collections::HashMap, io, sync::Arc};
 
 use super::util::DelayMap;
-use crate::base::middleware::{Action, ActionResponse, MiddlewareError};
+use crate::base::middleware::{Action, ActionResponse};
 use crate::base::{Buffer, Config, Package, Point, Stream, StreamStatus};
 
 #[derive(Error, Debug)]
@@ -29,8 +29,6 @@ pub enum Error {
     Codec(#[from] LinesCodecError),
     #[error("Serde error {0}")]
     Json(#[from] serde_json::error::Error),
-    #[error("Download OTA error")]
-    Actions(#[from] MiddlewareError),
     #[error("Couldn't fill stream")]
     Stream(#[from] crate::base::Error),
 }
